New sender component implementation, needs pulish
